+++ src/lib/components/VideoCard.svelte
@@ -1,34 +1,50 @@
 <script lang="ts">
     import { createEventDispatcher } from 'svelte'
     import { fade } from 'svelte/transition'
-    import { Heading } from '.'
+    import { Div, Heading } from '.'
+    import type { TransitionReceiver } from 'urfu-ui-kit'
 
+    export let id: string = undefined
     export let name = ''
+    export let node: HTMLElement = undefined
     export let position = ''
     export let description = ''
     export let src = ''
+    export let transition: TransitionReceiver = { in: undefined, out: undefined }
 
     let video: HTMLVideoElement
+    let videoPlaying = false
     const dispatch = createEventDispatcher()
 
-    let videoPlaying = false
     const toggleVideo = () => {
         if (src) {
             if (!videoPlaying) {
                 video.play()
                 videoPlaying = true
+                dispatch('play')
             }
             else {
                 video.pause()
                 videoPlaying = false
+                dispatch('pause')
             }
         }
         dispatch('click')
     }
 </script>
 
-<div class="video-card">
-    <div class="img" on:click={ toggleVideo }>
+<Div
+    {id}
+    bind:node
+    className="video-card"
+    {transition}
+    on:click={ toggleVideo }
+    on:mouseover
+    on:mouseleave
+    on:focus
+    on:blur
+>
+    <div class="img">
         <!-- svelte-ignore a11y-media-has-caption -->
         <video class="lazy" loop playsinline bind:this={ video } width="100%">
             <source {src} data-src={ src } type="video/mp4">
@@ -42,10 +58,10 @@
     { #if description }
         <p class="description">{ @html description }</p>
     { /if }
-</div>
+</Div>
 
 <style>
-    .video-card .img {
+    :global(.video-card[data-kit-component="true"] .img) {
         display: grid;
         position: relative;
         place-items: center;
@@ -57,32 +73,32 @@
         cursor: pointer;
     }
 
-    .video-card .img > * {
+    :global(.video-card[data-kit-component="true"] .img > *) {
         grid-row: 1;
         grid-column: 1;
     }
 
-    .video-card .img video {
+    :global(.video-card[data-kit-component="true"] .img video) {
         object-fit: cover;
         /* max-width: 239px; */
     }
 
-    .video-card .img img {
+    :global(.video-card[data-kit-component="true"] .img img) {
         transition: 0.1s ease-in-out;
         transform: scale(1);
     }
 
-    .video-card:hover .img img {
+    :global(.video-card[data-kit-component="true"]:hover .img img) {
         transform: scale(1.2);
     }
 
     @media screen and (max-width: 768px) {
-        .video-card .img {
+        :global(.video-card[data-kit-component="true"] .img) {
             min-width: 220px;
         }
     }
 
-    .video-card p.description {
+    :global(.video-card[data-kit-component="true"] p.description) {
         color: rgba(0, 0, 0, .6);
     }
 </style>