+++ src/routes/component/base/Input.svelte
@@ -40,7 +40,7 @@
         <Heading size={4} color="var(--blue)">wide <ParameterType value="boolean" /></Heading>
         <Text>Если параметр указан (имеет значение <code>true</code>), компонент будет иметь свойство <code>width: 100%</code></Text>
         <Heading size={4} color="var(--blue)">mask <ParameterType value="object" /></Heading>
-        <Text>Маска для поля ввода (Подробнее: <Link href="https://www.npmjs.com/package/svelte-imask" target="_BLANK">svelte-imask</Link>)</Text>
+        <Text>Маска для поля ввода (Подробнее: <Link href="https://www.npmjs.com/package/imask" target="_BLANK">svelte-imask</Link>)</Text>
         <Heading size={4} color="var(--blue)">id <ParameterType value="string" /></Heading>
         <Text>Аналог стандартного атрибута id</Text>
         <Heading size={4} color="var(--blue)">min <ParameterType value="number" /></Heading>