+++ src/routes/component/display/Partner.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { Code, Heading, Tag, Grid, Partner, Text, ParameterType } from '$lib/components'
+    import { Code, Heading, Tag, Grid, Link, Partner, Text, ParameterType } from '$lib/components'
     import CodeExample from '$lib/codes/PartnerExample'
 
     let partners = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]
@@ -33,6 +33,16 @@
         <Text>Ссылка на изображение</Text>
         <Heading size={4} color="var(--blue)">caption <ParameterType value="string" /></Heading>
         <Text>Подпись (опциональная)</Text>
+        <Heading size={4} color="var(--blue)">transition <ParameterType value="object" /></Heading>
+        <Text>
+            Объект, описывающий переходы для компонента. Имеет следующие свойства: <br />
+            <code>in</code> – переход при монтировании компонента <br />
+            <code>out</code> – переход при демонтировании компонента <br /> <br />
+            Каждое из этих свойств может содержать в себе объект со следующими свойствами: <br />
+            <code>func</code> – функция, выполняющая переход от начального состояния к конечному <br />
+            <code>options</code> – опциональный объект с опциональными свойствами <code>delay</code>, <code>duration</code> и <code>easing</code> <br />
+            Подробнее: <Link href="https://svelte.dev/docs#template-syntax-element-directives-transition-fn" target="_BLANK">Документация по Svelte</Link>
+        </Text>
         <Heading size={4} color="var(--blue)">node <ParameterType value="HTMLElement" /></Heading>
         <Text>Параметр для связки с HTML-элементом, который лежит в основе компонента <br /> (аналог стандартного <code>bind:this</code>)</Text>
         <Heading size={4} color="var(--blue)">id <ParameterType value="string" /></Heading>